+++ doc/source/code/tut_import.py
@@ -33,8 +33,8 @@ cellRule = {'label': 'PYR_HH3D_rule', 'conditions': {'cellType': 'PYR', 'cellMod
 utils.importCell(cellRule=cellRule, fileName='geom.hoc', cellName='E21')
 cellRule['sections']['soma']['mechs']['hh'] = {'gnabar': 0.12, 'gkbar': 0.036, 'gl': 0.003, 'el': -70}  		# soma hh mechanism
 for secName in cellRule['sections']:
-	cellRule['sections'][secName]['mechs']['pas'] = {'g': 0.0000357, 'e': -70}
-	cellRule['sections'][secName]['geom']['cm'] = 10
+ 	cellRule['sections'][secName]['mechs']['pas'] = {'g': 0.0000357, 'e': -70}
+ 	cellRule['sections'][secName]['geom']['cm'] = 1
 netParams['cellParams'].append(cellRule)  
 
 ### Traub
@@ -88,21 +88,14 @@ netParams['synMechParams'].append({'label': 'AMPA', 'mod': 'Exp2Syn', 'tau1': 1.
 netParams['connParams'] = []  
 
 netParams['connParams'].append({
-	'preTags': {'popLabel': 'background'}, 'postTags': {'cellType': 'PYR', 'cellModel': ['Traub', 'Mainen']}, # background -> PYR (weight=0.001)
-	'connFunc': 'fullConn', 	# connectivity function (all-to-all)
-	'weight': 0.001, 			# synaptic weight 
-	'delay': 5,					# transmission delay (ms) 
-	'sec': 'soma'})		
-
-netParams['connParams'].append({
-	'preTags': {'popLabel': 'background'}, 'postTags': {'cellType': 'PYR', 'cellModel': ['HH', 'HH3D', 'Izhi2003b', 'Izhi2007b']}, # background -> PYR (weight=0.1)
+	'preTags': {'popLabel': 'background'}, 'postTags': {'cellType': 'PYR', 'cellModel': ['Traub', 'HH', 'Mainen', 'Izhi2003b', 'Izhi2007b']}, # background -> PYR (weight=0.1)
 	'connFunc': 'fullConn', 	# connectivity function (all-to-all)
 	'weight': 0.1, 			# synaptic weight 
 	'delay': 5,					# transmission delay (ms) 
 	'sec': 'soma'})		
 
 netParams['connParams'].append({
-	'preTags': {'popLabel': 'background'}, 'postTags': {'cellType': 'PYR', 'cellModel': ['Friesen', 'Izhi2003a', 'Izhi2007a']}, # background -> PYR (weight = 10)
+	'preTags': {'popLabel': 'background'}, 'postTags': {'cellType': 'PYR', 'cellModel': ['HH3D','Friesen','Izhi2003a', 'Izhi2007a']}, # background -> PYR (weight = 10)
 	'connFunc': 'fullConn', 	# connectivity function (all-to-all)
 	'weight': 5, 				# synaptic weight 
 	'delay': 5,					# transmission delay (ms) 
@@ -127,8 +120,10 @@ simConfig['recordTraces'] = {'V_soma':{'sec':'soma','loc':0.5,'var':'v'}}  # Dic
 simConfig['recordStep'] = 1 			# Step size in ms to save data (eg. V traces, LFP, etc)
 simConfig['filename'] = 'model_output'  # Set file output name
 simConfig['savePickle'] = False 		# Save params, network and sim output to pickle file
-simConfig['plotRaster'] = True 			# Plot a raster
-simConfig['plotCells'] = [1] 			# Plot recorded traces for this list of cells
+
+simConfig['analysis'] = {}
+simConfig['analysis']['plotRaster'] = {'orderInverse': True}			# Plot a raster
+simConfig['analysis']['plotTraces'] = {'include': [0]} 			# Plot recorded traces for this list of cells
 
 
 # Create network and run simulation